// src/api/model/tileHealth.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 0.0.8
 */
import type { TileType } from './tileType';
import type { TileSlice } from './tileSlice';

export interface TileHealth {
  type: TileType;
  startDate: Date;
  endDate: Date;
  expected: number;
  actual: number;
  slices: TileSlice[];
}